Updated to use swish-protocol 2.2.1
Updated to use swish-protocol 2.2.1 which has several major revisions. Name of HandshakeClient was changed to SwishClient. New interface SwishPackage which the project now utilizes.

Request and request-promise now moved to the dependencie (previously accidentally saved in the devDependency

disabled "no-param-reassign": " in eslint as it warns on trivial changes we need to make to a parame's attribute